requirements.txt
accelerate==1.11.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.42.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
elevenlabs==2.23.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.12.0
networkx==3.4.2
numpy==1.25.2
openai==2.8.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
protobuf==4.25.8
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
rich==13.9.4
rpds-py==0.29.0
safetensors==0.6.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.35.0
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.22.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.100
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.0
websockets==15.0.1
